style: split files
