+++ pkg/controller/generate-registry/generate.go
@@ -5,13 +5,11 @@ import (
 	"fmt"
 	"io"
 	"os"
-	"reflect"
 	"strings"
 
 	"github.com/aquaproj/aqua/pkg/config"
 	"github.com/aquaproj/aqua/pkg/config/registry"
 	"github.com/aquaproj/aqua/pkg/github"
-	"github.com/mholt/archiver/v3"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/afero"
 	"gopkg.in/yaml.v3"
@@ -254,161 +252,6 @@ func (ctrl *Controller) parseAssetInfos(pkgInfo *registry.PackageInfo, assetInfo
 	ctrl.setSupportedEnvs(envs, pkgInfo)
 }
 
-func (ctrl *Controller) setSupportedEnvs(envs map[string]struct{}, pkgInfo *registry.PackageInfo) { //nolint:cyclop
-	if has(envs, "darwin") || has(envs, "darwin/amd64") {
-		pkgInfo.SupportedEnvs = append(pkgInfo.SupportedEnvs, "darwin")
-	} else if has(envs, "darwin/arm64") {
-		pkgInfo.SupportedEnvs = append(pkgInfo.SupportedEnvs, "darwin/arm64")
-	}
-	if has(envs, "darwin/amd64") && !has(envs, "darwin/arm64") {
-		pkgInfo.Rosetta2 = boolP(true)
-	}
-	if has(envs, "linux/amd64") {
-		if has(envs, "linux/arm64") {
-			pkgInfo.SupportedEnvs = append(pkgInfo.SupportedEnvs, "linux")
-		} else {
-			pkgInfo.SupportedEnvs = append(pkgInfo.SupportedEnvs, "linux/amd64")
-		}
-	} else if has(envs, "linux/arm64") {
-		pkgInfo.SupportedEnvs = append(pkgInfo.SupportedEnvs, "linux/arm64")
-	}
-	if has(envs, "windows/amd64") {
-		if has(envs, "windows/arm64") {
-			pkgInfo.SupportedEnvs = append(pkgInfo.SupportedEnvs, "windows")
-		} else {
-			pkgInfo.SupportedEnvs = append(pkgInfo.SupportedEnvs, "windows/amd64")
-		}
-	} else if has(envs, "windows/arm64") {
-		pkgInfo.SupportedEnvs = append(pkgInfo.SupportedEnvs, "windows/arm64")
-	}
-	if reflect.DeepEqual(pkgInfo.SupportedEnvs, []string{"darwin", "linux", "windows"}) {
-		pkgInfo.SupportedEnvs = nil
-	}
-	if reflect.DeepEqual(pkgInfo.SupportedEnvs, []string{"darwin", "linux", "windows/amd64"}) {
-		pkgInfo.SupportedEnvs = []string{"darwin", "linux", "amd64"}
-	}
-	if reflect.DeepEqual(pkgInfo.SupportedEnvs, []string{"darwin", "linux/amd64", "windows/amd64"}) {
-		pkgInfo.SupportedEnvs = []string{"darwin", "amd64"}
-	}
-}
-
-func (ctrl *Controller) setOS(assetName, lowAssetName string, assetInfo *AssetInfo) { //nolint:funlen
-	osList := []*OS{
-		{
-			Name: "apple-darwin",
-			OS:   "darwin",
-		},
-		{
-			Name: "unknown-linux-gnu",
-			OS:   "linux",
-		},
-		{
-			Name: "unknown-linux-musl",
-			OS:   "linux",
-		},
-		{
-			Name: "pc-windows-msvc",
-			OS:   "windows",
-		},
-		{
-			Name: "pc-windows-gnu",
-			OS:   "windows",
-		},
-		{
-			Name: "darwin",
-			OS:   "darwin",
-		},
-		{
-			Name: "linux",
-			OS:   "linux",
-		},
-		{
-			Name: "windows",
-			OS:   "windows",
-		},
-		{
-			Name: "apple",
-			OS:   "darwin",
-		},
-		{
-			Name: "osx",
-			OS:   "darwin",
-		},
-		{
-			Name: "macos",
-			OS:   "darwin",
-		},
-		{
-			Name: "mac",
-			OS:   "darwin",
-		},
-		{
-			Name: "win64",
-			OS:   "windows",
-		},
-		{
-			Name: "win",
-			OS:   "windows",
-		},
-	}
-	for _, o := range osList {
-		if idx := strings.Index(lowAssetName, o.Name); idx != -1 {
-			osName := assetName[idx : idx+len(o.Name)]
-			assetInfo.OS = o.OS
-			if osName != o.OS {
-				if assetInfo.Replacements == nil {
-					assetInfo.Replacements = map[string]string{}
-				}
-				assetInfo.Replacements[o.OS] = osName
-			}
-			assetInfo.Template = strings.Replace(assetInfo.Template, osName, "{{.OS}}", 1)
-			break
-		}
-	}
-	if assetInfo.OS == "" && strings.Contains(lowAssetName, ".exe") {
-		assetInfo.OS = "windows"
-	}
-}
-
-func (ctrl *Controller) setArch(assetName, lowAssetName string, assetInfo *AssetInfo) {
-	archList := []*Arch{
-		{
-			Name: "amd64",
-			Arch: "amd64",
-		},
-		{
-			Name: "arm64",
-			Arch: "arm64",
-		},
-		{
-			Name: "x86_64",
-			Arch: "amd64",
-		},
-		{
-			Name: "64bit",
-			Arch: "amd64",
-		},
-		{
-			Name: "aarch64",
-			Arch: "arm64",
-		},
-	}
-	for _, o := range archList {
-		if idx := strings.Index(lowAssetName, o.Name); idx != -1 {
-			archName := assetName[idx : idx+len(o.Name)]
-			assetInfo.Arch = o.Arch
-			if archName != o.Arch {
-				if assetInfo.Replacements == nil {
-					assetInfo.Replacements = map[string]string{}
-				}
-				assetInfo.Replacements[o.Arch] = archName
-			}
-			assetInfo.Template = strings.Replace(assetInfo.Template, archName, "{{.Arch}}", 1)
-			break
-		}
-	}
-}
-
 const formatRaw = "raw"
 
 func (ctrl *Controller) parseAssetName(assetName, version string) *AssetInfo {
@@ -428,65 +271,4 @@ func (ctrl *Controller) parseAssetName(assetName, version string) *AssetInfo {
 	}
 	assetInfo.Format = ctrl.getFormat(assetName)
 	return assetInfo
-}
-
-func (ctrl *Controller) getFormat(assetName string) string { //nolint:funlen,cyclop
-	a, err := archiver.ByExtension(assetName)
-	if err != nil {
-		return formatRaw
-	}
-	switch a.(type) {
-	case *archiver.Rar:
-		return "rar"
-	case *archiver.Tar:
-		return "tar"
-	case *archiver.TarBrotli:
-		if strings.HasSuffix(assetName, ".tbr") {
-			return "tbr"
-		}
-		return "tar.br"
-	case *archiver.TarBz2:
-		if strings.HasSuffix(assetName, ".tbz2") {
-			return "btz2"
-		}
-		return "tar.bz2"
-	case *archiver.TarGz:
-		if strings.HasSuffix(assetName, ".tgz") {
-			return "tgz"
-		}
-		return "tar.gz"
-	case *archiver.TarLz4:
-		if strings.HasSuffix(assetName, ".tlz4") {
-			return "tlz4"
-		}
-		return "tar.lz4"
-	case *archiver.TarSz:
-		if strings.HasSuffix(assetName, ".tsz") {
-			return "tsz"
-		}
-		return "tar.sz"
-	case *archiver.TarXz:
-		if strings.HasSuffix(assetName, ".txz") {
-			return "txz"
-		}
-		return "tar.xz"
-	case *archiver.TarZstd:
-		return "tar.zsd"
-	case *archiver.Zip:
-		return "zip"
-	case *archiver.Gz:
-		return "gz"
-	case *archiver.Bz2:
-		return "bz2"
-	case *archiver.Lz4:
-		return "lz4"
-	case *archiver.Snappy:
-		return "sz"
-	case *archiver.Xz:
-		return "xz"
-	case *archiver.Zstd:
-		return "zst"
-	default:
-		return formatRaw
-	}
 }
